requirements.txt
aiogram==3.16.0
python-dotenv==1.0.1
shapely
requests
matplotlib
pandas